Move create_response to libapi
